refactor: delete independent SiteTitle
