+++ packages/gatsby-theme-project-portal/src/components/Navbar.tsx
@@ -1,9 +1,7 @@
 import React, { FunctionComponent } from "react"
 import { Link } from "gatsby"
 import { FaBars, FaTimes } from "react-icons/fa"
-import { ImageDataLike } from "gatsby-plugin-image"
-import { SiteTitle, NavbarItem } from "."
-
+import { GatsbyImage, getImage, ImageDataLike } from "gatsby-plugin-image"
 export interface NavbarProps {
   title?: string
   activePage?: string
@@ -15,8 +13,6 @@ export interface NavbarProps {
   }[]
 }
 
-import { SiteTitle } from "."
-
 export const Navbar: FunctionComponent<NavbarProps> = ({
   title,
   activePage,
@@ -24,6 +20,7 @@ export const Navbar: FunctionComponent<NavbarProps> = ({
   pages,
 }: NavbarProps) => {
   const [navbarOpen, setNavbarOpen] = React.useState(false)
+  const resolvedImage = getImage(image)
   return (
     <>
       <nav
@@ -46,7 +43,16 @@ export const Navbar: FunctionComponent<NavbarProps> = ({
               className="block xl:min-h-full mx-4 my-3 xl:my-auto overflow-hidden text-nav text-black font-bold flex gap-4 items-center whitespace-nowrap"
               to="/"
             >
-              <SiteTitle image={image} title={title} />
+              {resolvedImage ? (
+                <GatsbyImage
+                  className="xl:inline-block logotype"
+                  image={resolvedImage}
+                  alt={"nav_logo"}
+                />
+              ) : (
+                <></>
+              )}
+              <div>{title}</div>
             </Link>
           </div>
           <div
