--- packages/gatsby-theme-project-portal/src/components/SiteTitle.stories.tsx
@@ -1,39 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react"
-import { SiteTitle } from "./SiteTitle"
-
-import { loadImage } from "./Story.utilities"
-
-const meta: Meta<typeof SiteTitle> = {
-  component: SiteTitle,
-  tags: ["autodocs"],
-  includeStories: /^[A-Z]/,
-}
-
-export const siteTitleLogo = loadImage("image/logo-square.png", 64, 74, "fixed")
-
-export default meta
-
-type Story = StoryObj<typeof SiteTitle>
-
-export const Primary: Story = {
-  args: {
-    image: siteTitleLogo,
-    title: "The Site Title",
-  },
-}
-
-export const Nulls: Story = {
-  args: {
-    image: null,
-    title: null,
-  },
-}
-
-export const EmptyString: Story = {
-  args: {
-    image: null,
-    title: "",
-  },
-}
-
-export const NoArgs: Story = {}
