+++ packages/gatsby-theme-project-portal/src/components/index.tsx
@@ -23,7 +23,6 @@ export type { ProjectDetailProps } from "./ProjectDetail"
 export { ProjectPage } from "./ProjectPage"
 export type { ProjectPageProps } from "./ProjectPage"
 export type { TopicType } from "./Topics"
-export { SiteTitle } from "./SiteTitle"
 export type { SiteTitleProps } from "./SiteTitle"
 export { SearchBar } from "./SearchBar"
 export { Topics } from "./Topics"
